Add main_shared Makefile target, create main_shared.c out of main.c library use example
